requirements.txt
# Core Framework
streamlit>=1.28.0
python-dotenv>=1.0.0

# LangChain for RAG
langchain>=0.0.340
langchain-community>=0.0.10
langchain-core>=0.1.3

# Local LLM Support (Ollama)
ollama>=0.1.0

# Document Processing
pypdf2>=3.0.0
python-docx>=0.8.11
python-pptx>=0.6.21

# Text Processing & Thai Language Support
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0
numpy>=1.21.0
pandas>=2.0.0

# Vector Database
faiss-cpu>=1.7.0
chromadb>=0.4.0

# Thai Language Processing
pythainlp>=4.0.0

# Utilities
tqdm>=4.60.0
requests>=2.25.0
pydantic>=2.0.0

# File Processing
chardet>=5.0.0